# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wedding Invitation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [business]
      summary: Promote a Buyer to Business Account or return them to Buyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated account role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRoleResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /rsvp:
    get:
      operationId: listRsvps
      tags: [rsvp]
      summary: List all RSVPs
      responses:
        "200":
          description: List of RSVPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rsvp"
    post:
      operationId: createRsvp
      tags: [rsvp]
      summary: Submit RSVP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRsvpBody"
      responses:
        "201":
          description: RSVP created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rsvp"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rsvp/count:
    get:
      operationId: getRsvpCount
      tags: [rsvp]
      summary: Get RSVP count summary
      parameters:
        - name: invitationToken
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: RSVP counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RsvpCount"
...
components:
  schemas:
...
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum:
            - buyer
            - business_account
      required:
        - role
    UserRoleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - buyer
            - business_account
      required:
        - id
        - name
        - email
        - role
    Rsvp:
      type: object
      properties:
        id:
          type: integer
        invitationToken:
          type: string
        name:
          type: string
        attending:
          type: boolean
        numberOfGuests:
          type: integer
        timeSlot:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required:
        - id
        - invitationToken
        - name
        - attending
        - numberOfGuests
        - createdAt
    CreateRsvpBody:
      type: object
      properties:
        invitationToken:
          type: string
        name:
          type: string
        attending:
          type: boolean
        numberOfGuests:
          type: integer
        timeSlot:
          type: string
        message:
          type: string
      required:
        - invitationToken
        - name
        - attending
        - numberOfGuests
    RsvpCount:
      type: object
      properties:
        attending:
          type: integer
        notAttending:
          type: integer
        totalGuests:
          type: integer
      required:
        - attending
        - notAttending
        - totalGuests
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error
...